# frontend/generator-project/swagger/auth.json
{
  "swagger": "2.0",
  "info": {
    "title": "Auth BFF Service - Terms and Conditions",
    "description": "API for user authentication and terms and conditions management.",
    "version": "1.0.0"
  },
  "host": "localhost:8087",
  "basePath": "/",
  "schemes": [
    "http"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "securityDefinitions": {
    "bearerAuth": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header",
      "description": "JWT Bearer token"
    }
  },
  "tags": [
    {
      "name": "Auth",
      "description": "User authentication and terms & conditions"
    }
  ],
  "paths": {
    "/rest/api/v1/auth/login": {
      "post": {
        "summary": "Authenticate user",
        "description": "Authenticate user with email and password. Returns session token and user profile.",
        "operationId": "login",
        "tags": [
          "Auth"
        ],
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/LoginRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Login success",
            "schema": {
              "allOf": [
                {
                  "$ref": "#/definitions/APIResponse"
                },
                {
                  "type": "object",
                  "properties": {
                    "data": {
                      "$ref": "#/definitions/LoginResponse"
                    }
                  }
                }
              ]
            }
          },
          "400": {
            "description": "Missing or malformed fields",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          },
          "401": {
            "description": "Invalid credentials",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          },
          "500": {
            "description": "Internal server error",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          }
        }
      }
    },
    "/rest/api/v1/auth/terms": {
      "get": {
        "summary": "Retrieve terms and conditions",
        "description": "Retrieve terms and conditions content.",
        "operationId": "getTerms",
        "tags": [
          "Auth"
        ],
        "responses": {
          "200": {
            "description": "Terms content object",
            "schema": {
              "allOf": [
                {
                  "$ref": "#/definitions/APIResponse"
                },
                {
                  "type": "object",
                  "properties": {
                    "data": {
                      "$ref": "#/definitions/TermsContentResponse"
                    }
                  }
                }
              ]
            }
          },
          "500": {
            "description": "Server error",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          }
        }
      }
    },
    "/rest/api/v1/auth/terms/accept": {
      "post": {
        "summary": "Accept terms and conditions",
        "description": "Record that the authenticated user has accepted the terms.",
        "operationId": "acceptTerms",
        "tags": [
          "Auth"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/TermsAcceptRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Acceptance recorded",
            "schema": {
              "$ref": "#/definitions/APIResponse"
            }
          },
          "400": {
            "description": "Missing fields",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          },
          "401": {
            "description": "Unauthenticated",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          },
          "500": {
            "description": "Internal server error",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          }
        }
      }
    }
  },
  "definitions": {
    "LoginRequest": {
      "type": "object",
      "required": [
        "userId",
        "password"
      ],
      "properties": {
        "userId": {
          "type": "string",
          "example": "user123"
        },
        "password": {
          "type": "string",
          "format": "password",
          "example": "password123"
        }
      }
    },
    "LoginResponse": {
      "type": "object",
      "properties": {
        "token": {
          "type": "string",
          "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
        },
        "expiresAt": {
          "type": "string",
          "format": "date-time",
          "example": "2026-05-06T15:00:00Z"
        },
        "user": {
          "$ref": "#/definitions/UserResponse"
        }
      }
    },
    "UserResponse": {
      "type": "object",
      "properties": {
        "userId": {
          "type": "string",
          "example": "user123"
        },
        "fullName": {
          "type": "string",
          "example": "John Doe"
        },
        "role": {
          "type": "string",
          "example": "ADMIN"
        },
        "branchCode": {
          "type": "string",
          "example": "001"
        },
        "branchName": {
          "type": "string",
          "example": "Head Office"
        }
      }
    },
    "TermsContentResponse": {
      "type": "object",
      "properties": {
        "title": {
          "type": "string",
          "example": "Terms and Conditions"
        },
        "lastUpdated": {
          "type": "string",
          "format": "date-time",
          "example": "2026-05-01T00:00:00Z"
        },
        "sections": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/TermsSection"
          }
        }
      }
    },
    "TermsSection": {
      "type": "object",
      "properties": {
        "heading": {
          "type": "string",
          "example": "1. Introduction"
        },
        "body": {
          "type": "string",
          "example": "Welcome to our service..."
        }
      }
    },
    "TermsAcceptRequest": {
      "type": "object",
      "required": [
        "acceptedAt"
      ],
      "properties": {
        "acceptedAt": {
          "type": "string",
          "format": "date-time",
          "description": "ISO8601 timestamp when the user accepted the terms.",
          "example": "2026-05-06T10:30:00Z"
        }
      }
    },
    "APIResponse": {
      "type": "object",
      "properties": {
        "status": {
          "type": "string",
          "enum": [
            "success",
            "error"
          ],
          "example": "success"
        },
        "data": {
          "type": "object"
        },
        "message": {
          "type": "string",
          "example": "OK"
        }
      }
    },
    "ErrorResponse": {
      "type": "object",
      "properties": {
        "status": {
          "type": "string",
          "example": "error"
        },
        "message": {
          "type": "string",
          "example": "Invalid credentials"
        }
      }
    }
  }
}